cryptohome: Break up auth factor IsSupported
This splits the IsSupported function up into IsSupportedByHardware and
IsSupportedByStorage pieces.  The motivation here is that there are some
contexts where I want to be able to do the hardware checks only, which
is difficult with the current API where you either have to go out and
look up the other factors in use or come up with some kind of stub
values for the parameters. Breaking the function into two pieces makes
this much clearer.

A side benefit is that this also lets us define a more reusable
implementation for most of the IsSupportedByStorage checks, instead of
having to duplicate things like the "no kiosk factors" check into every
implementation.

BUG=b:282052310
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I5825899d13fe667e5988a43681e22256bc06566a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4521734
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>