vm_tools: Add a blockid to external disks
Concierge both shares and mounts external disks into its vms on-request.
This is problematic because some vms (bru, reference) mount the disk
themselves which causes an error.

Eventually we will remove concierge's mount call, but to get there we
need to first make it easy for the guests to identify the external drive
themselves, hence the id.

BUG=b:280685257
TEST=emerge, check /sys/block/vd*/serial in guest

Change-Id: If9cfc16883bd052d0fdd22ef19154b4f23d3984f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4520605
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: James Ye <jamesye@google.com>