cryptohome: Send AuthFactor back on successful Add/update
This CL appends AuthFactor to the end of AddAuthFactorReply and
UpdateAuthFactorReply. I decided to put the changes in userdataauth.cc
instead of auth_session.cc because Update And Authenticate follow
similar path for USS Migration and we are not adding any information for
Authenticate. This change is the simplest change we could make.

BUG=b:279741905
TEST=CQ Passes

Change-Id: I2ad41bf1d010fdc09877a387f002e4357e927886
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4526102
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>