hiberman: Check if metrics file exists before opening it
On a regular boot without resume from hibernate the metrics file
doesn't exist. Currently we try to open the metrics file and log
an error if it doesn't exist. The error is misleading when the
file isn't expected to exist. Before opening the file check if
it exists, if it doesn't return and log a debug message.

BUG=b:232184089
TEST=no error log on regular boot

Change-Id: I6127ca44a00a3b2d879dfc86abbf232b40948c17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4528423
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>