metrics: Exempt usb_pd_device project to always be recorded to the disk
Allow usb_pd_device project to always be recorded to the disk
disregarding the UMA consent status. This is due to the known flakiness
where structured metrics may not be properly saved to the disk despite
UMA consent is on. Considering high priority of usb_pd_device metrics,
add it to exemption so that we can collect metrics for sure. This
exemption shall be removed once the flakiness is fixed.

If UMA consent is off, this bypass would save usb_pd_device metrics to
the disk, but Chromium Metrics Reader would not report it to Google
servers. The saved metrics are removed from the disk after 10 minutes.

BUG=b:281887744
TEST=emerge-${BOARD} metrics

Change-Id: I161eac5a46d8004532107571ff0b21b74253759f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4528715
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Won Chung <wonchung@google.com>
Tested-by: Won Chung <wonchung@google.com>
Reviewed-by: Tony Yeoman <tby@chromium.org>