cryptohome: data_migrator: Generalize chunk size calculation
Currently, cryptohome::data_migrator::MigrationHelper uses the size of
the free disk space of the migration destination to calculate the
migration chunk size for files. However, in ARCVM /data migration, the
size of the free disk space of the migration destination is not the
actual amount of space that is available for the migrator.
This CL allows the delegate class of MigrationHelper to specify how to
get the free disk space size for the migrator, and change it in
ArcVmDataMigrationHelperDelegate to use the size of the free space on
the stateful partition, which should be the actual amount of free space
for arcvm-data-migrator.

BUG=b:280248293
TEST=FEATURES=test emerge-hatch cryptohome
TEST=FEATURES=test emerge-hatch arcvm-data-migrator
TEST=tast run <DUT> arc.VMDataMigration.*
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v*

Change-Id: I0e4d5628dc24f54110677cbbd88bc2a609346107
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4530017
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>