patchpanel: fix the presubmit check "check-source-files-included"
Recently, "check-source-files-included" presubmit check is introduced
into platform2 that checks the modified source files are included in
BUILD.gn. However, the source files of patchpanel are listed in the
array variables, instead of the GN build target. So the presubmit
check failed when changing those files.

This CL moves the list of files to the GN build target to fix the
presubmit check.

BUG=b:273373137
TEST=FEATURES=test emerge-brya patchpanel
TEST=the following CL passes the presubmit check

Change-Id: Ie00aba3b0733de18dbfb68cda7ccdf190561e297
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4529682
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>