vm_tools: cros_im: Exclude qtbase-abi in cros-im depends
The qtbase-abi package is incompatible with Ubuntu users, and is
not a dependency since we already depend on libqt5core5a.

This CL excludes this package from being added via
shlib:Depends. For details on dh_shlibdeps, see
https://man7.org/linux/man-pages/man1/dh_shlibdeps.1.html

BUG=b:282131093
TEST=kokoro, manually build cros-im and apt-cache show cros-im does
not include qtbase-abi as a depends

Change-Id: I82b4f6a7a1cd0c35daaa9d7b6d14642a14081b0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4530329
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Sophia Lin <sophialin@google.com>