lorgnette: Enumerate USB devices
The first step of detecting scanners is to enumerate all the USB
devices.  This lets lorgnette directly add entries for IPP-USB scanners
that support eSCL and gives a way to track which potential devices may
require a DLC download.  This CL implements the basic libusb enumeration
loop.  Probing detected IPP-USB devices will come in the following CL.

Because libusb itself isn't directly testable, this also includes a set
of wrappers and fakes that let tests inject whatever USB devices they
want.  In this CL, this is used to test the device prober logic.  Once
the implementation is complete enough to send d-bus signals, this can be
used to cover the main public API as well.

BUG=b:277049537
TEST=unit tests; lorgnette_cli discover

Change-Id: Ide57f6be8fa3cc9e05215eb8705266bc8d116f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4534081
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>