dns-proxy: Controller shutdown waits for child processes
DNS proxy tests rely on restarting the service. However, child processes
of DNS proxy might not stop right away if the service is being stopped.
One example of such case is when there is a queue of tasks for the child
processes.

In order to have a reliable restart, the controller needs to wait for
its child processes to be properly stopped. This is done by monitoring
the child processes pid with waitpid().

BUG=b:281778714
TEST=tast run <DUT> network.DNSProxy* network.ResolvConf*
TEST=restart dns-proxy; observe that the service is re-started only
     after all child processes are stopped;

Change-Id: I2bc5608d4c33ab48c2be019ac3c237e187144b23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4536415
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tested-by: Jason Iman <jasongustaman@chromium.org>