cryptohome: Refactor the flag file check into a function
This pulls the mechanics of checking for flag files in
/var/lib/cryptohome into a reusable function. In particular, pull the
knowledge of where the file lives into filesystem_layout instead of it
being hard-coded into the code doing a check.

We plan to use this pattern again in other code so this should make it
slightly easier to enforce consistency.

BUG=b:282006449
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib952f775e427673e315b5e174cd4c761b82c33c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4539335
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>