cryptohome: Add flag file to enable fingerprint login
This adds in support for enabling using the fingerprint factor for full
decrypt by adding a file to flag it. This is modelled after the pattern
used for the USS enable/disable files.

In this case only an enable file is added, because we don't intend for
there to be any situation where login is enabled by default and then
disabled via the flag file. Once the ability to adjust the fingerprint
policy via API is added any tests using the flag file should be
converted over to using the API.

BUG=b:282006449
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Iba39cb25c6e4d8a8f762f89c5ff330f1f9875fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4540841
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>