metrics: Add nonblocking lock option to WriteMetricsTofile
Currently crash_reporter has a risk of deadlocking when calling
Record*ToUMA if the crashed process held the lock on
/var/lib/metrics/uma-events.

Add an optional parameter to the SynchronousMetricsWriter constructor to
fail writing the metric if the lock is unavailable after two attempts so
we don't deadlock. If the lock is unavailable, metrics_library will
sleep for a short duration before trying again. This is a best effort
attempt to record the metric instead of dropping it if the lock is
initially unavailable.

BUG=b:270732207, b:232987294
TEST=unit tests

Fixed: b:232987294
Change-Id: If87591dc78e1b27ed620288e36a6f9da5e4c7471
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4549191
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Kendrake Tsui <kendraketsui@google.com>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>