libbrillo: Use transparent comparator in VariantDictionary
Transparent comparator will allow lookups with keys whose types are
comparable with std::string, e.g., base::StringPiece and
std::string_view. Without transparent comparator, we have to construct
a string at first, which is not always necessary.

Note that in most of the current use cases of VariantDictionary, the
keys are static which are usually c-style strings, skipping the
unnecessary construction of strings will help a lot in both runtime
performance and the generated code size. For example, changing
the interface of shill::KeyValueStrore (which is a wrapper of this
class) to use base::StringPiece instead of `const std::string&` reduces
the binary size of shill by ~40KB.

BUG=b:283037672
TEST=FEATURES="test" emerge-${BOARD} libbrillo

Change-Id: I591918d396f97addae7e587041b6c9c45565f5eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4542223
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>