shill: cellular: tethering network selection logic in the Cellular object
The logic to select which Network to use for tethering is complex, and
depends not only on capabilities of the cellular device being used,
but also on the current operator settings as specified in the mobile
operators database.

The service provider will request the tethering specific Network to
the cellular device explicitly, regardless of whether a new connection
needs to be brought up or an existing one reused.

The only supported usecase so far is the one where we reuse the
Default connection also for Tethering.

BUG=b:249388816
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: I010207251f4d81c17b1b16860fd3ae7c5812d2af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4570859
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>