hiberman: Make resume time calculation more robust
Return a duration of zero when UNIX_EPOCH.elapsed() returns
an error instead of crashing. The function would return an
error in the unlikely case that the system time is < UNIX_EPOCH.

Also account for the case that the system time moves backwards
between the start and finishing of the resume process. Report
a resume duration of 0s if that happens.

BUG=b:232184089
TEST=manual

Change-Id: I5805260c32d96d58d46576e0906f053d80329b73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4543926
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>