init: clobber_state: switch to RedirectOutputToMemory for output
This uses O_TMPFILE to guarantee the output is created and is simpler
than relying on the base tempfile API which uses mkstemp (and more).
Since we never need the actual file to exist, just access to the output
of the tool, switch over to this API.

Memory-wise it's the same -- both use the same underlying tempdir, and
that is backed by tmpfs.

This also provides a little bit better logging when something goes
wrong.  We can include the name of the failing process rather than a
temporary path with no meaning.

BUG=None
TEST=CQ passes

Change-Id: I680138fd2cd2f43d6351ce2beb4f39b271961b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4543880
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>