arc: Support false positive cases of ReferencesParent in data migrator
Currently, libchrome's base::FilePath::ReferencesParent() can return
true for paths with no references to their parents (via "..") if they
have components like "...", " ..", and ".. ".

This behavior was introduced to deal with the way Windows handles path
components but in a platform-agnostic way for simplicity: crbug/181617.

Although in general it would not be so problematic to be strict against
unusual paths, an issue is that it causes ARCVM /data migration to fail
if there is such a file in the migration source: b/280247852.

Ideally it should be fixed in libchrome, but it seems to be difficult as
the function is referenced from various places: See crbug/464760.

As a compromise solution, this CL implements a workaround by defining
arc::data_migrator::Platform, which is a derived class of
cryptohome::Platform. The class just overrides InitializeFile(), which
is the problematic function that calls ReferencesParent().

The overriding version first calls the original version. If it fails in
such a way that it is possibly due to ReferencesParent(), it falls back
to another code path. There, almost the same thing as the original is
performed, but instead of calling base::FilePath::ReferencesParent(), it
just checks whether the given path contains ".." as its component.

This redundant structure is intentionally chosen to confine the effect
of the workaround to a small set of cases where the migration is doomed
to fail otherwise. This is because the implementation of the workaround
duplicates a function in libchrome (base::File::Initialize()), but not
necessarily catches up to its latest version. On the other hand, the
original version directly calls the libchrome function, and thus it is
preferred as long as it does not fail at ReferencesParent().
This redundancy should not cost much because it is almost equivalent to
calling the same function twice when a fairly rare type of file path is
found.

Since this solution is far from ideal, the class should be removed once
the aforementioned issue is resolved, or it is decided to be practically
useless.

BUG=b:280247852
TEST=FEATURES=test emerge-${BOARD} arcvm-data-migrator
TEST=This CL makes arc.VMDataMigration.* with CL:4547631 FAIL > PASS.

Change-Id: I98a405532ba3ed387161976fb65d9ffc2e6dcd0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4547106
Reviewed-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>