diagnostics: Add crash_sender invoker to executor
Add an invoker of crash_sender to executor. To be used by crash events.

BUG=b:266018436
TEST=FEATURES=test emerge-$BOARD diagnostics
TEST=Manually verified that crash_sender is invoked properly with
CL:4382221. It uses the same test as CL:4382221. After installing the
modified cros_healthd, the test was done right after reboot and after a
user log out to ensure that the mount of `/home/chronos` happens before
cros_healthd starts and it won't be remounted.

Change-Id: I2b3b24a9598220fd4f3eb50f34696551ba99e55c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4548108
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Hong Xu <xuhong@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Hong Xu <xuhong@google.com>