camera: common: Refactor camera metadata deserialization
This CL refactors the metadata deserialization by using the `entries`
field as the source of truth, and validates the self consistency for
each entry.

As noted in http://b/215648588, the unused fields would be deprecated
and removed in the future. There would be another series of CLs for
cleanup the usage on other components, such as Chromium and ARC.

BUG=b:281872677,b:281780998,b:281761871,b:215648588
TEST=Pass camera.CCAUISmoke.vivid on Betty.

Change-Id: I9c06ddfe1b49d33e5d7845ec10aa90ba91b9b473
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4547193
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Shik Chen <shik@chromium.org>
Auto-Submit: Shik Chen <shik@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>