cryptohome: Add CLI to get AuthFactorStatusUpdate
This CLI will add the capability to fetch AuthFactorStatusUpdate
signal. This would be purely used to fetch the consequent signals
that are sent after the user is locked out. The first signal is
sent when the lockout attempt occurs which could be checked with
|DoAuthenticationWithStatusUpdate|. But afterwards signals are
sent periodically and the new CLI will add the capability to fetch
each of those by creating a run_loop to wait for the next signal.

BUG=b:260904713
TEST=Manual test of CLI

Change-Id: Ic80cef5d5be3f2e9efc129f57e235c229ab4da1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4549188
Tested-by: Behnood Momenzadeh <behnoodm@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>