patchpanel: shill_client: maintain Device properties
This patch adds to ShillClient a map of Device indexed by shill Device
DBus path where ShillClient::Device properties are updated continuously
on notifications from shill when:
  - a physical Device is added or removed
  - properties of a physical Device are changed (most properties are
    ignored by patchpanel)

This map will replace internally in ShillClient the map of network
interfaces and IPConfig objects.

VPN connections are not tracked with this mechanism since the Device of
a connected VPN is already tracked through the logical default device.

The comments in ShillClient are updated to reflect which methods and
which callbacks take into account VPN connections and which methods and
callbacks only consider physical networks.

BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Ibca391651efc0b1dd3f2ea75410d5edb7a1dc471
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4548731
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>