cryptohome: cap fuzzed paths at 2*PATH_MAX
We're running into some timeouts on fuzz tests because you can get
arbitrary strings used as paths which seems to be particularly
antagonistic for base::FilePath::GetComponents. In the test case in
particular I've been debugging, you get a path with ~16k components
which takes over two minutes for FilePath to split.

I could try and pin down the performance problems, or try rewriting code
to avoid GetComponents, but in practice it doesn't seem like we really
need the fuzzer to be exercising enormous paths? That's not really the
point of the test here.

BUG=b:279284073
TEST=cros_fuzz with the test case in the linked bug

Change-Id: Ic08f91d9b89f3262e88c57a55e75d666cbc45e01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4556535
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>