missive: Fix race in MissiveArgs
The flakiness seems to be caused by the fact that
`FeatureLib::ListenForRefetchNeeded` makes a call to `attached_callback`
asynchronously, while `MissiveArgs` was reporting the initial parameters
settings without waiting for that.
As a results, tests could encounter a race - started to wait for the
parameters update before the listener was attached.

BUG=b:283981044
TEST=FEATURES=test emerge-${BOARD} missive

Change-Id: I4cdd53d1e6efe2bc627e1ed9cc761ed799daa47d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4559727
Reviewed-by: Josh Hilke <jrhilke@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>