vm_tools: Move shared library installation from debian to meson scripts
To install cros-im, shared libraries built by meson should be copied to
the relevant /usr/lib folders. Previously this was done using the debian
build scripts, but this CL moves this copy into the meson build
scripts. Doing so ensures that the libraries are installed not only when
the cros-im deb package is installed, but also when a developer runs
'meson install' command manually.

BUG=b:253980498
TEST=manually run the new readme instructions or ./build_packages.sh amd64 and
check that the resultant deb installs correctly.

Change-Id: I214e07ab254a0d2205713cd762bfa26b97882631
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4595417
Tested-by: Sophia Lin <sophialin@google.com>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Sophia Lin <sophialin@google.com>