Update print and scan OWNERS
For every directory that has a DIR_METADATA file pointing to the Paper
I/O team, point the matching OWNERS file to our shared OWNERS file.
This updates several projects that were previously using cups_proxy as a
shared list and cleans up the (mostly outdated) lists of individual
OWNERS.

The subdirectories of system_api are left as-is because they point to
the code they're logically part of instead of the shared file.

BUG=b:234625517
TEST=CQ

Change-Id: I1d8484f11bec076d26645b1c329932b18580f1a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4562826
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>