permission_broker: Allow Google Twinkie V2
Google Twinkie V2 is a new USB PD analyzer which should work with GUIs
in crostini. Currently crostini is unable to access Google Twinkie V2
because it gets denied by the permission broker. This CL adds Google
Twinkie V2 to the kSerialAllowedIds list for it to be accessible by
crostini.

BUG=b:284187067
TEST=FEATURES=test emerge-${BOARD} chromeos-base/permission_broker

Change-Id: Ia86bc82112f8a03bb31752d764bc2f70ed07f44b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4563258
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>