power: add support for Charge Limit
This adds support for Charge Limit, which lets the system prevent
charging over the Adaptive Charging hold percent at all times, to
powerd. The hold percent is 80% for the default value, but it's
configurable (even though we don't have a Chrome setting to change it).

BUG=b:280880357
TEST=unittests

Change-Id: I343fc944ae11aab4ac7591fa09b0edfd07b68c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4568354
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>