cryptohome: Create a small interface class for UserMetadata
This class is basically a small wrapper around the USS classes that can
give you a way to access UserMetadata in a neat little package.

Ideally we could make some improvements under the covers, for example
adding some way to take advantage of in-memory copies of the data when
it's loaded into an AuthSession rather than always hitting the on-disk
data, but this is a basic reliable version that makes it easier to
access this data without having to pass around raw access to the USS
storage to everyone.

BUG=b:284301590
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Iba8b5e20686be66fb951441ead3d1071f041b059
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4568344
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Howard Yang <hcyang@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>