patchpanel: change the offset meaning of Subnet
Originally, the offset argument of Subnet means the difference between
the first usable host address, which is (base address + 1). The
decision only causes confusion and doesn't benefit anything.

This CL changes the reference point of offset to the base address of
the Subnet. With that, we could avoid off-by-one handling inside
Subnet class.

BUG=b:283369026
TEST=FEATURES=test emerge-brya patchpanel vm_host_tools

Change-Id: I739a11b15232d4fd11ab9b834f5aea2808cf55ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4569627
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>