hiberman: Don´t expect a read error to indicate the end of the snapshot image
SnapshotDevice::load_image() currently expects a read error
to indicate the end of the snapshot image, due to invalid
integrity data beyond the image. That was a valid assumption
as long as only one hibernation cycle was supported. With
multiple cycles there could be valid integrity data beyond
the image, from a larger image from an earlier cycle.

Don't expect the error and interpret a write of 0 bytes as
EOF conditions. Still abort loading the image when a read
error is encountered and the subsequent write doesn't
indicate EOF.

BUG=b:284376414
TEST=multiple hibernate cycles with varying memory usage succeed

Change-Id: If05c70de051690ff2ce5194aff815e48ebbd8931
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4573875
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>