diagnostics: Change cpu_cache to use V2 adapter in diagnostics service
Change the implementation of cpu_cache in diagnostics service to use
V2 adapter.

After this change, the CPU cache routine will attempt to run
stressapptest but leaving at least 500MiB of memory for the system.
stressapptest needs to run with at least 128MiB, so a total of 628MiB
is needed for the routine to run.

If the routine does not have over 628MiB memory, the routine will
have a status of kError.

BUG=b:182964589
TEST=(cr) FEATURES=test USE="-cros-debug" emerge-volteer diagnostics
     (dut) cros-health-tool diag --action=run_routine --routine=cpu_cache

Change-Id: I8fdfc14f0be0ef614b753207386545e1940b032c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4570007
Reviewed-by: Denny Huang <dennyh@google.com>
Commit-Queue: Ethan Cheng <yycheng@google.com>
Tested-by: Ethan Cheng <yycheng@google.com>