patchpanel: Subnet return IPv4CIDR with offset
Originally, Subnet::AddressAtOffset() return a uint32_t to represent
the IPv4 address which offset by the Subnet's base address. But most
of case what we want is to return a IPv4 address with offset and the
same prefix length as Subnet.

This CL changes the method to return a net_base::IPv4CIDR to express
the intention of the method. Also, it uses std::optional<T> to
indicate the situation that the offset is out of range. It could force
the caller to check the failure case.

BUG=b:283369026
TEST=FEATURES=test emerge-brya patchpanel vm_host_tools

Change-Id: Iff928f152b928c994da05084217e246c5dce9333
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4567778
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>