shill: cellular: skip APN try list build in capability
The capability logic takes care of setting up a PDN connection in
ModemManager, and will now receive the APN try list in the Connect()
call itself, instead of building it based on the selected APN type.

This allows us to attempt a PDN connection of type DEFAULT with the
tethering-specific APN try list, for example.

BUG=b:283396208
TEST=cros_workon_make --board=$BOARD --test shill

Change-Id: Id4841b384f624d99592bca31f851921fbe2df554
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4573693
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>