shill: Decouple NetworkPriority from Connection implementation details
This patch uses a generic ranking_order in NetworkPriority to describe
the order of the networks (0 for default, 1 for secondary, etc) instead
of the ip rule priority, which is a Connection implementation detail.

Connection now calculates the ip rule priority from this ranking_order
internally.

BUG=b:264963034
TEST=unit
TEST=tast run $DUT network.Routing*

Change-Id: I9cc2a9a9771685fc4ab5aeb1d3c7aec6ebfd47f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4563536
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>