vm_tools: refvm: Enable installation of refvm
This makes the reference VM self-installing. It can be booted
interactively or as an installer, based on the presence of the SMBIOS
OEM string "refvm:install=true".

The LVM VG name can be specified during image building. A new VG name is
generated when installed.

BUG=b:284092752
TEST=Install refvm with Bruschetta installer

Change-Id: I0b940d0fa23202cf8e41e0d6ca09acb0382d646f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4577104
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: James Ye <jamesye@google.com>