resourced: Add support for resourced-lib / libresourced.so
The resourced runs inside host OS and uses gRPC framework to get/set
the platform related parameters. Resourced implements both gRPC client
and server. The other side also needs to implement gRPC client and
server. The libresourceD.so implements the gRPC related code for the
other side and exports the APIs. The main binary which wants to
communicate with resourced running on the host side can load this
library dynamically.

This patch adds the support for building libresourceD.so. The code for
it will be maintained in platform2/resourced/interface. This feature
will be controlled through 'lib_resourced' flag which will be disabled
by default.

BUG=None
TEST=Verified borealis VM with application communicating with libresourced.so
in guest which in turn communicates with the resourced on the host OS.

Co-authored-by: Vaibhav Vyas <vavyas@nvidia.com>
Co-authored-by: Abhishek Sahu <abhsahu@nvidia.com>

Change-Id: Ia0717d32ec346e1adf6e0dd69f4af4aecc1a0229
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4553231
Auto-Submit: Vaibhav Vyas (SW) US <vavyas@nvidia.com>
Commit-Queue: Vaibhav Vyas (SW) US <vavyas@nvidia.com>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Vaibhav Vyas (SW) US <vavyas@nvidia.com>