shill: Apply both IPv4 and IPv6 DNS setting into resolv.conf
Previously, Network only create resolv.conf using the IPConfig that was
used to SetupConnection - that is, IPv4 if IPv4 is available, and IPv6
if the network is IPv6 only.

This patch changes the behavior so that DNS configuration from both IPv4
and IPv6 are applied.

BUG=b:264963034
BUG=b:265680125
TEST=unit

Change-Id: I3c827a0aba3f37711b58fd5c8af21c106b4c994d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4573208
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>