cryptohome: Clean up uses of unique_ptr reset
This replaces a bunch of calls of the form reset(new X(...)) with the
more modern =std::unique_ptr<X>(...). In general it's preferred to avoid
naked new as much as possible, and so this type of code tends to be
flagged by linter or editor warnings.

I also cleaned up a few other types of warnings (mostly removing
redundant declarations and definitions) if they just happened to be
adjacent to the code I was editing. Low hanging fruit.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ifa27e3086ed8159e02e614eb63da28eed1730876
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4572718
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>