shill: Use base::StringPiece in PropertyStore (part 1)
This series of patches replace the `const std::string&` with
`base::StringPiece` in the parameters of the functions related to
PropertyStore. All the patches reduce shill's binary size by about 6KB
in total on octopus-kernelnext. The size reduction mainly happens in the
ctors where properties are registers, e.g., the ctor size of Manager
changes from ~8KB to ~6KB.

This patch replaces the function parameters on the PropertyStore class.
Since the functions of PropertyStore do not pass the input string
parameter to other functions, this change has almost no side effect,
while avoids creating the temporary string objects when the function is
called with string literals.

BUG=b:284259970
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: I4c9ca7f63075a5cf1f52ba7dc28f523a6d07e420
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4576384
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>