shill: Distinguish disconnection metrics from attempt failure
The |ConnectionAttemptResult| event in structured metrics for a failure
connection attempt should be mutually exclusive with the |Disconnection|
event by design. However, evidence in b/284044303 shows they are
actually not and under edge cases both events can be emitted at the same
time. This CL fixes this issue.

|ConnectionAttemptResult| for a failure connection attempt is emitted
when a service is being disconnected and |is_attempt_failure| is true.
That is, the device is trying to connect with a BSS but failed. On the
contrary, |Disconnection| means disconnection from a connected BSS and
is handled by a different pathway. It is emitted when shill receives the
signal from wpa_supplicant that updates the disconnection reason code.

However, because wpa_supplicant does not distinguish if a disconnection
is associated with a connection attempt or an established connection,
wpa_supplicant may update the disconnection reason code during an
attempt, which makes shill emit both events. For example, the connection
attempt failure during 4-way handshake stage is accompanied by reason
code 15. Therefore, the two events are not guaranteed to be mutually
exclusive. Instead, the |Disconnection| event should only be emitted
when the disconnection is not associated with an attempt.

Here, we distinguish whether the reason code update is associated with
an attempt failure or not where shill emits |Disconnection|:
  - Before: Emit anyway when the disconnection reason code is updated;
  - After: Only emit |Disconnection| when |is_attempt_failure| is false.

The same logic to obtain |is_attempt_failure| is used for both events.
After this change, these two events become (nearly) two sides of a coin
controlled by the boolean |is_attempt_failure|. This guarantees that
they will be mutually exclusive with our best effort.

BUG=b:284044303
TEST=FEATURES=test emerge-${BOARD} shill
TEST=manual test (also see b/254314211 comment#20):
  1. Set metrics.cc SHILL_LOG_LEVEL=-2
  2. Find a PSK network, enter a wrong passphrase.
  3. Check net.log and make sure only the |ConnectionAttemptResult|
  event is sent and the |Disconnection| event is not sent.

Change-Id: I23a265ac5fa5072e5d61ac189a9636ff381c7a41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4577400
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Yichen Yu <yichenyu@google.com>
Commit-Queue: Yichen Yu <yichenyu@google.com>
Auto-Submit: Yichen Yu <yichenyu@google.com>